+++ src/main/java/empresaTransporte/utp/repositorio/AreaRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AreaRepository extends JpaRepository<Area, Integer> {
+public interface AreaRepository extends JpaRepository<Area, Long> {
 
     List<Area> findAreaByActivoTrue();
