+++ src/main/java/empresaTransporte/utp/repositorio/UbigeoDistRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UbigeoDistRepository extends JpaRepository<UbigeoDist, Integer> {
+public interface UbigeoDistRepository extends JpaRepository<UbigeoDist, Long> {
 
     List<UbigeoDist> findUbigeoDistByActivoTrue();
